% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{as_kde}
\alias{as_kde}
\title{Convert list object to density class}
\usage{
as_kde(object, ngrid, ...)
}
\arguments{
\item{object}{List with components \code{y} (a matrix of points) and \code{density}
(the value of the density at each row of \code{y})}

\item{ngrid}{Number of points to use for the grid in each dimension. Default is
10001 for univariate densities and 101 for multivariate densities.}

\item{...}{Additional arguments are ignored.}
}
\value{
An object of class "kde"
}
\description{
Convert list object to density class
}
\examples{
y <- seq(-4, 4, by = 0.01)
as_kde(list(y = y, density = dnorm(y)))
}
\author{
Rob J Hyndman
}
