% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autolayer.R
\name{autolayer.distribution}
\alias{autolayer.distribution}
\title{Add ggplot layer of densities from distributional objects in 1 dimension}
\usage{
\method{autolayer}{distribution}(object, ngrid = 501, ...)
}
\arguments{
\item{object}{distribution object from the distributional package or
\code{\link{dist_kde}}()}

\item{ngrid}{Number of points at which to evaluate the density function.}

\item{...}{Additional arguments are passed to \code{\link[ggplot2]{geom_line}}.}
}
\value{
A ggplot layer
}
\description{
Add ggplot layer of densities from distributional objects in 1 dimension
}
\details{
This function adds a ggplot layer of a density from a distributional object.
For univariate densities, it adds a line plot of the density function.
For bivariate densities, it adds a contour plot of the density function.
}
\examples{
# Univariate density
dist_kde(c(rnorm(500), rnorm(500, 4, 1.5))) |>
  autoplot(show_hdr = TRUE, prob = c(0.5, 0.95), color = "#c14b14") +
  autolayer(distributional::dist_normal(), linetype = "dashed")
}
\author{
Rob J Hyndman
}
