% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bagplot.R
\name{gg_bagplot}
\alias{gg_bagplot}
\title{Bagplot}
\usage{
gg_bagplot(data, var1, var2, col = "#0072B2", ...)
}
\arguments{
\item{data}{A data frame or matrix containing the data.}

\item{var1}{The name of the first variable to plot (a bare expression).}

\item{var2}{The name of the second variable to plot (a bare expression).}

\item{col}{The main color to use. The median is shown at full strength, while the
bag and loop have some transparency added.}

\item{...}{Other arguments currently ignored.}
}
\value{
A ggplot object showing a bagplot of the data.
}
\description{
Produces a bivariate bagplot. A bagplot is analagous to a
univariate boxplot, except it is in two dimensions. Like a boxplot, it
shows the median, a region containing 50\% of the observations, a region
showing the remaining observations other than outliers, and any outliers.
}
\examples{
gg_bagplot(n01, v1, v2)
}
\references{
Rousseeuw, P. J., Ruts, I., & Tukey, J. W. (1999).
The bagplot: A bivariate boxplot. The American Statistician, 52(4), 382–387.
}
\seealso{
\code{\link[aplpack]{bagplot}}
}
\author{
Rob J Hyndman
}
