% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{density.matrix}
\alias{density.matrix}
\alias{density.data.frame}
\alias{density.numeric}
\title{Kernel density estimation}
\usage{
\method{density}{matrix}(x, h, H, ...)

\method{density}{data.frame}(x, h, H, ...)

\method{density}{numeric}(x, h, ...)
}
\arguments{
\item{x}{A numeric vector or matrix of data values.}

\item{h}{A numeric vector of bandwidths for univariate data. The default is
obtained using \code{\link[ks]{hns}}.}

\item{H}{A numeric matrix of bandwidths for multivariate data. The default is
obtained using \code{\link[ks]{Hns}}.}

\item{...}{Additional arguments are passed to \code{\link[ks]{kde}}.}
}
\description{
Kernel density estimation for multivariate data up to 6 dimensions. The
bandwidths are chosen to be suitable for tail estimation. The calculations
are handled by the \code{\link[ks]{kde}} function, but using different default
bandwidths.
}
\examples{
# Univariate density
y <- c(rnorm(100), rnorm(100, 3, 1))
density(y) |>
  autoplot() +
  geom_rug(data = tibble(y=y), aes(x=y))
# Bivariate density
tibble(y1 = rnorm(100), y2 = y1 + rnorm(100)) |>
  density() |>
  autoplot(filled = TRUE)
}
