% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_density.R
\name{dist_density}
\alias{dist_density}
\title{Create distributional object based on a specified density}
\usage{
dist_density(x, density)
}
\arguments{
\item{x}{Numerical vector of ordinates, or a list of such vectors.}

\item{density}{Numerical vector of density values, or a list of such vectors.}
}
\description{
Creates a distributional object using a density specified as pair of vectors
giving (x, f(x)). The density is assumed to be piecewise linear between the
points provided, and 0 outside the range of x.
}
\examples{
dist_density(seq(-4, 4, by = 0.01), dnorm(seq(-4, 4, by = 0.01)))

}
