% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air_quality.R
\name{fetch_air_quality}
\alias{fetch_air_quality}
\alias{air_quality}
\title{Air quality data for 12 Beijing monitoring stations from 2013 to 2017}
\format{
A data frame with 420,768 rows and 17 columns:
\describe{
\item{year}{Year of measurement}
\item{month}{Month of measurement}
\item{day}{Day of measurement}
\item{hour}{Hour of measurement (0--23)}
\item{pm2_5}{Particulate matter with diameter less than 2.5 micrometers
(micrograms per cubic meter)}
\item{pm10}{Particulate matter with diameter less than 10 micrometers
(micrograms per cubic meter)}
\item{so2}{Sulfur dioxide concentration (micrograms per cubic meter)}
\item{no2}{Nitrogen dioxide concentration (micrograms per cubic meter)}
\item{co}{Carbon monoxide concentration (micrograms per cubic meter)}
\item{o3}{Ozone concentration (micrograms per cubic meter)}
\item{temperature}{Temperature (degrees Celsius)}
\item{pressure}{Atmospheric pressure (hPa)}
\item{dew_point}{Dew point temperature (degrees Celsius)}
\item{rainfall}{Rainfall (millimeters)}
\item{wind_direction}{Wind direction}
\item{wind_speed}{Wind speed (meters per second)}
\item{station}{Name of the monitoring station}
}
}
\source{
UCI Machine Learning Repository
\url{https://archive.ics.uci.edu/dataset/501/beijing+multi+site+air+quality+data}
}
\usage{
fetch_air_quality()
}
\value{
Data frame
}
\description{
Hourly air quality measurements from 12 monitoring stations across Beijing,
China, from 1 March 2013 to 28 February 2017. The data are downloaded and
returned.
}
\examples{
\dontrun{
air_quality <- fetch_air_quality()
air_quality |>
  filter(station == "Aotizhongxin") |>
  ggplot(aes(x = temperature, y = pm2_5)) +
  geom_point(alpha = 0.1)
}
}
\references{
Hyndman, R J (2026) \emph{That's weird: Anomaly detection using R},
\url{https://OTexts.com/weird/}.
}
