% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{fr_mortality}
\alias{fr_mortality}
\title{French mortality rates by age and sex}
\format{
A data frame with 31,648 rows and 4 columns.
}
\source{
Human Mortality Database \url{https://www.mortality.org}
}
\usage{
fr_mortality
}
\value{
Data frame
}
\description{
A data set containing French mortality rates between the years 1816 and 1999,
by age and sex.
}
\examples{
fr_mortality
}
\references{
Rob J Hyndman (2026) "That's weird: Anomaly detection using R", Section 1.4,
\url{https://OTexts.com/weird/}.
}
\keyword{datasets}
