% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_density.R
\name{gg_density}
\alias{gg_density}
\title{Produce ggplot of densities from distributional objects in 1 or 2 dimensions}
\usage{
gg_density(
  object,
  prob = seq(9)/10,
  fill = FALSE,
  show_hdr = FALSE,
  show_points = FALSE,
  show_mode = FALSE,
  show_lookout = FALSE,
  ngrid = 501,
  color = "#00659e",
  alpha = NULL,
  jitter = FALSE,
  ...
)
}
\arguments{
\item{object}{distribution object from the distributional package or
\code{\link{dist_kde}}()}

\item{prob}{Probability of the HDR regions to be drawn.}

\item{fill}{If \code{TRUE}, and the density is bivariate, the bivariate contours
are shown as filled regions rather than lines.}

\item{show_hdr}{If \code{TRUE}, and the density is univariate, then the HDR regions
specified by \code{prob} are shown as a ribbon below the density.}

\item{show_points}{If \code{TRUE}, then individual observations are plotted.}

\item{show_mode}{If \code{TRUE}, then the mode of the distribution is shown as a
point.}

\item{show_lookout}{If \code{TRUE}, then the observations with lookout probabilities less than 0.05 are shown in red.}

\item{ngrid}{Number of points at which to evaluate the density function.}

\item{color}{Color used for mode and HDR contours/regions when there is only
one distribution. Otherwise a ggplot color scale is used.}

\item{alpha}{Transparency of points. Ignored if \code{show_points} is \code{FALSE}.
Defaults to min(1, 500/n), where n is the number of observations plotted.}

\item{jitter}{When TRUE and \code{show_points} is TRUE, a small amount of vertical
jittering is applied to the observations for a univariate distribution.}

\item{...}{Additional arguments are currently ignored.}
}
\value{
A ggplot object.
}
\description{
Produce ggplot of densities from distributional objects in 1 or 2 dimensions
}
\details{
This function produces a ggplot of a density from a distributional object.
For univariate densities, it produces a line plot of the density function, with
an optional ribbon showing some highest density regions (HDRs) and/or the observations.
For bivariate densities, it produces a contour plot of the density function, with
the observations optionally shown as points.
The mode can also be drawn as a point.
For bivariate densities, the combination of fill = TRUE\verb{, }show_points = TRUE\verb{, }show_mode = TRUE\verb{, and }prob = c(0.5, 0.99)\verb{is equivalent to showing an HDR boxplot beneath the density plot. For univariate densities,  the combination of}show_hdr = TRUE\verb{, }show_points = TRUE\verb{, }show_mode = TRUE\verb{, and }prob = c(0.5, 0.99)` is equivalent to an HDR boxplot.
}
\examples{
# Univariate density
kde <- dist_kde(c(rnorm(500), rnorm(500, 4, .5)))
gg_density(kde,
  show_hdr = TRUE, prob = c(0.5, 0.95), color = "#c14b14",
  show_mode = TRUE, show_points = TRUE, jitter = TRUE
)
c(dist_normal(), kde) |>
  gg_density(show_hdr = TRUE, prob = c(0.5, 0.95))
tibble(y1 = rnorm(5000), y2 = y1 + rnorm(5000)) |>
  dist_kde() |>
  gg_density(show_points = TRUE, alpha = 0.1, fill = TRUE)
}
\author{
Rob J Hyndman
}
