% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_dist.R
\name{gg_dist}
\alias{gg_dist}
\title{Produce ggplot of densities from distributional objects in 1 or 2 dimensions}
\usage{
gg_dist(
  object,
  prob = seq(9)/10,
  fill = FALSE,
  show_hdr = FALSE,
  show_points = FALSE,
  show_mode = FALSE,
  show_lookout = FALSE,
  ngrid = 501,
  color = "#00659e",
  palette = hdr_palette,
  alpha = NULL,
  jitter = FALSE,
  ...
)
}
\arguments{
\item{object}{distribution object from the distributional package or
\code{\link{dist_kde}}()}

\item{prob}{Probability of the HDR contours to be drawn (for a bivariate plot only).}

\item{fill}{If \code{TRUE}, and the density is bivariate, the bivariate contours
are shown as filled regions rather than lines.}

\item{show_hdr}{If \code{TRUE}, and the density is univariate, then the HDR regions
specified by \code{prob} are shown as a ribbon below the density.}

\item{show_points}{If \code{TRUE}, then individual observations are plotted.}

\item{show_mode}{If \code{TRUE}, then the mode of the distribution is shown.}

\item{show_lookout}{If \code{TRUE}, then the observations with lookout probabilities less than 0.05 are shown in red.}

\item{ngrid}{Number of points at which to evaluate the density function.}

\item{color}{Color used for mode and HDR contours. If \code{palette = hdr_palette},
it is also used as the basis for HDR regions.}

\item{palette}{Color palette function to use for HDR filled regions
(if \code{fill} is \code{TRUE} or \code{show_hdr} is \code{TRUE}).}

\item{alpha}{Transparency of points. When \code{fill} is \code{FALSE}, defaults to
min(1, 1000/n), where n is the number of observations. Otherwise, set to 1.}

\item{jitter}{When TRUE and \code{show_points} is TRUE, a small amount of vertical
jittering is applied to the observations.}

\item{...}{Additional arguments are currently ignored.}
}
\value{
A ggplot object.
}
\description{
Produce ggplot of densities from distributional objects in 1 or 2 dimensions
}
\details{
This function produces a ggplot of a density from a distributional object.
For univariate densities, it produces a line plot of the density function, with
an optional ribbon showing some highest density regions (HDRs) and/or the observations.
For bivariate densities, it produces a contour plot of the density function, with
the observations optionally shown as points.
The mode can also be drawn as a point with the HDRs.
For bivariate densities, the combination of \code{fill = TRUE}, \code{show_points = TRUE},
\code{show_mode = TRUE}, and \code{prob = c(0.5, 0.99)} is equivalent to an HDR boxplot.
For univariate densities,  the combination of \code{show_hdr = TRUE}, \code{show_points = TRUE},
\code{show_mode = TRUE}, and \code{prob = c(0.5, 0.99)} is equivalent to an HDR boxplot.
}
\examples{
# Univariate density
dist_kde(c(rnorm(500), rnorm(500, 4, 1.5))) |>
  gg_dist(show_hdr = TRUE, prob = c(0.5, 0.95), color = "#c14b14")
ymat <- tibble(y1 = rnorm(5000), y2 = y1 + rnorm(5000))
#ymat |>
#  dist_kde(ymat) |>
#  gg_dist(show_points = TRUE, alpha = 0.1, fill = TRUE)
}
\author{
Rob J Hyndman
}
