% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmm.R
\name{gmm_to_dist}
\alias{gmm_to_dist}
\title{Convert Gaussian mixture model to a distributional object}
\usage{
gmm_to_dist(object)
}
\arguments{
\item{object}{Object of class \code{Mclust}, output from the [\link[mclust:Mclust]{mclust::Mclust}] function}
}
\value{
An object of class \code{distributional}
}
\description{
Convert Gaussian mixture model to a distributional object
}
\examples{
\dontrun{
library(mclust)
# Create a bivariate Gaussian mixture model for the Old Faithful data
gmm <- oldfaithful |>
  select(duration, waiting) |>
  Mclust() |>
  gmm_to_dist()
gg_density(gmm) +
  geom_point(data = oldfaithful, aes(x = duration, y = waiting), alpha = 0.1) +
  labs(x = "Duration", y = "Waiting time")
}
}
