% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grubbs.R
\name{grubbs_anomalies}
\alias{grubbs_anomalies}
\title{Statistical tests for anomalies using Grubbs' test and Dixon's test}
\usage{
grubbs_anomalies(y, alpha = 0.05)
}
\arguments{
\item{y}{numerical vector of observations}

\item{alpha}{size of the test.}
}
\description{
Grubbs' test (proposed in 1950) identifies possible anomalies in univariate
data using z-scores assuming the data come from a normal distribution.
}
\details{
The test is based on z-scores, and a point is identified as an
anomaly when the associated absolute z-score is greater than a threshold value.
A vector of logical values is returned, where \code{TRUE} indicates an anomaly.
This version of Grubbs' test looks for outliers anywhere in the sample.
Grubbs' original test came in several variations which looked for one outlier,
or two outliers in one tail, or two outliers on opposite tails. These variations
are implemented in the \code{\link[outliers]{grubbs.test}} function.
}
\examples{
y <- c(rnorm(1000), 3:6)
tibble(y = y) |> filter(grubbs_anomalies(y))
}
\references{
Grubbs, F. E. (1950). Sample criteria for testing outlying observations.
\emph{Annals of Mathematical Statistics}, 21(1), 27–58.
}
\seealso{
\code{\link[outliers]{grubbs.test}}, \code{\link[outliers]{dixons.test}}
}
\author{
Rob J Hyndman
}
