% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{gun_deaths}
\alias{gun_deaths}
\title{Gun ownership and homicide rates by country}
\format{
A data frame with 77 rows and
4 columns:
\describe{
\item{country}{Country name}
\item{gun_ownership_rate}{Gun ownership rate (number of guns owned by civilians per 100 people)}
\item{homicide_rate}{Homicide rate (number of homicides per 100,000 people where the weapon was a firearm)}
\item{region}{World region according to Our World in Data}
}
}
\source{
World Population Review \url{https://worldpopulationreview.com/country-rankings/gun-ownership-by-country} and \url{https://ourworldindata.org/grapher/homicide-rates-from-firearms}
}
\usage{
gun_deaths
}
\value{
Data frame
}
\description{
A data set containing gun ownership rates and homicide rates for 2017 for
various countries around the world. The gun ownership rates are the number of guns owned by civilians per 100 people. The homicide rates are the number of homicides per 100,000 people where the weapon was a firearm.
}
\examples{
gun_deaths
}
\keyword{datasets}
