% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hampel.R
\name{hampel}
\alias{hampel}
\title{Hampel filter}
\usage{
hampel(y, bandwidth, k = 3)
}
\arguments{
\item{y}{numeric vector containing time series}

\item{bandwidth}{integer width of the window around each observation}

\item{k}{numeric number of standard deviations to declare an outlier}
}
\value{
numeric vector with outliers replaced by the median of the window.
}
\description{
The Hampel filter is designed to find outliers in time series data using
mean absolute deviations in the vicinity of each observation.
}
\details{
The median absolute deviation is done in a window of times within
\code{bandwidth} of each observation. A point is declared an outlier if its MAD
value is more than k standard deviations
}
\examples{
set.seed(1)
tibble(
  time = seq(41),
  y = c(rnorm(20), 5, rnorm(20))
) |>
  mutate(ystar = hampel(y, bandwidth = 3, k = 4)) |>
  ggplot(aes(x = time, y = y)) +
  geom_line(color = "red") +
  geom_line(aes(y = ystar))
}
\author{
Rob J Hyndman
}
