% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{hdr_palette}
\alias{hdr_palette}
\title{Color palette designed for plotting Highest Density Regions}
\usage{
hdr_palette(n, color = "#00659e")
}
\arguments{
\item{n}{Number of colors to generate}

\item{color}{First color of vector.}
}
\value{
A function that takes a number \code{n} and returns a vector of \code{n} colors
}
\description{
A sequential color palette is returned, with the first color being \code{color}, and the rest of the colors
being a mix of \code{color} with increasing amounts of white.
}
