% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdr.R
\name{gg_hdrboxplot}
\alias{gg_hdrboxplot}
\title{HDR plot}
\usage{
gg_hdrboxplot(
  data,
  var1,
  var2 = NULL,
  prob = c(0.5, 0.99),
  col = "#00659e",
  scatterplot = FALSE,
  ...
)
}
\arguments{
\item{data}{A data frame or matrix containing the data.}

\item{var1}{The name of the first variable to plot (a bare expression).}

\item{var2}{Optionally, the name of the second variable to plot (a bare expression).}

\item{prob}{A numeric vector specifying the coverage probabilities for the HDRs.}

\item{col}{The base color to use for the smallest highest density region. Other colors
are generated by whitening this color.}

\item{scatterplot}{A logical argument indicating if a regular HDR plot is required
(\code{FALSE}), or if a scatterplot in the same colors is required (\code{TRUE}).}

\item{...}{Other arguments passed to \code{\link[ks]{kde}}.}
}
\value{
A ggplot object showing an HDR plot or scatterplot of the data.
}
\description{
Produces a 1d or 2d box plot of HDR regions.
}
\examples{
gg_hdrboxplot(n01, v1)
gg_hdrboxplot(n01, v1, v2, scatterplot = TRUE)
}
\references{
Hyndman, R J. (1996) Computing and Graphing Highest Density Regions,
\emph{The American Statistician}, \bold{50}(2), 120–126.
}
\author{
Rob J Hyndman
}
