% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde_bandwidth.R
\name{kde_bandwidth}
\alias{kde_bandwidth}
\title{Robust bandwidth estimation for kernel density estimation}
\usage{
kde_bandwidth(
  data,
  method = c("robust", "normal", "plugin", "scv", "lookout"),
  ...
)
}
\arguments{
\item{data}{A numeric matrix or data frame.}

\item{method}{A character string giving the method to use. Possibilities are:
\code{"normal"} (normal reference rule),
\code{"robust"} (a robust version of the normal reference rule, the default),
\code{"plugin"} (a plugin estimator),
\code{"scv"} (a smoothed cross-validation estimator), and
\code{"lookout"} (the bandwidth matrix estimate of Hyndman, Kandanaarachchi & Turner, 2026).}

\item{...}{Additional arguments are ignored.}
}
\value{
A matrix of bandwidths (or a scalar in the case of univariate data).
}
\description{
Bandwidth matrices are estimated in several ways including
a normal reference rule,
a robust version of the normal reference rule (default),
a plugin estimator,
a smoothed cross-validation estimator,
or using the approach of Hyndman, Kandanaarachchi & Turner (2026).
Details of each method are given in Hyndman (2026).
}
\examples{
# Univariate bandwidth calculation
kde_bandwidth(oldfaithful$duration)
# Bivariate bandwidth calculation
kde_bandwidth(oldfaithful[, c("duration", "waiting")])
}
\references{
Hyndman, R J, Kandanaarachchi, S & Turner, K (2026)
"When lookout sees crackle: Anomaly detection via kernel density estimation",
unpublished. \url{https://robjhyndman.com/publications/lookout2.html}

Hyndman, R J (2026) "That's weird: Anomaly detection using R", Section 2.9 and 3.9,
\url{https://OTexts.com/weird/}.
}
\author{
Rob J Hyndman
}
