% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde_scores.R
\name{kde_scores}
\alias{kde_scores}
\title{Kernel density estimate scores}
\usage{
kde_scores(y, h = stats::bw.nrd(y), binned = FALSE, loo = FALSE)
}
\arguments{
\item{y}{Numerical vector of data}

\item{h}{Bandwidth. Dfault: \code{\link[stats]{bw.nrd}(y)}}

\item{binned}{Should binning be used to speed up calculation? Default: FALSE}

\item{loo}{Should leave-one-scores be returned? Default: FALSE}
}
\value{
Numerical vector containing kde scores
}
\description{
Compute kernel density estimate scores,
defined as minus the log of the kernel density estimate at each observation.
A Gaussian kernel is used, with default h given by a normal reference rule.
}
\examples{
kde_scores(c(rnorm(9),5))
}
\seealso{
\code{\link[stats]{bandwidth}}
\code{\link[ks]{kde}}
}
\author{
Rob J Hyndman
}
