% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookout.R
\name{lookout_prob}
\alias{lookout_prob}
\title{Lookout probabilities}
\usage{
lookout_prob(object, ...)
}
\arguments{
\item{object}{A numerical data set.}

\item{...}{Other arguments are passed to \code{\link[lookout]{lookout}}.}
}
\value{
A numerical vector containing the lookout probabilities
}
\description{
The lookout algorithm (Kandanaarachchi & Hyndman, 2022) computes
leave-one-out surprisal probabilities from a kernel density estimate using a
Generalized Pareto distribution. The kernel density estimate uses a
bandwidth based on topological data analysis and a quadratic kernel. So it is
similar but not identical to using \code{\link{surprisals}} with \code{loo = TRUE}
and \code{approximation = "gdp"}. A low probability indicates a likely anomaly.
}
\examples{
# Univariate data
tibble(
  y = c(5, rnorm(49)),
  lookout = lookout_prob(y)
)
# Bivariate data
tibble(
  x = rnorm(50),
  y = c(5, rnorm(49)),
  lookout = lookout_prob(cbind(x, y))
)
# Using a regression model
fit_of <- lm(waiting ~ duration, data = oldfaithful)
broom::augment(fit_of) |>
  mutate(lookout = lookout_prob(.std.resid)) |>
  arrange(lookout)
}
\references{
Sevvandi Kandanaarachchi & Rob J Hyndman (2022) "Leave-one-out
kernel density estimates for outlier detection", \emph{J Computational & Graphical
Statistics}, \strong{31}(2), 586-599. \url{https://robjhyndman.com/publications/lookout/}
}
\seealso{
\code{\link[lookout]{lookout}}
}
\author{
Rob J Hyndman
}
