% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{n01}
\alias{n01}
\title{Multivariate standard normal data}
\format{
A data frame with 1000 rows and 10 columns.
}
\usage{
n01
}
\value{
Data frame
}
\description{
A synthetic data set containing 1000 observations on 10 variables generated
from independent standard normal distributions.
}
\examples{
n01
}
\keyword{datasets}
