% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{oldfaithful}
\alias{oldfaithful}
\title{Old faithful eruption data}
\format{
A data frame with 1937 rows and 3 columns:
\describe{
\item{time}{Time eruption started}
\item{duration}{Duration of eruption in seconds}
\item{waiting}{Time to the following eruption}
}
}
\source{
\url{http://geysertimes.org}
}
\usage{
oldfaithful
}
\description{
A data set containing data on recorded eruptions of Old Faithful since 2015.
Recordings are incomplete, especially during the winter months when observers
may not be present.
}
\keyword{datasets}
