% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weird.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{autoplot}
\alias{augment}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{broom}{\code{\link[broom:augment]{augment}}}

  \item{ggplot2}{\code{\link[ggplot2:autoplot]{autoplot()}}}
}}

