% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surprisals.R
\name{surprisals}
\alias{surprisals}
\alias{surprisals_prob}
\title{Surprisals and surprisal probabilities}
\usage{
surprisals(object, ...)

surprisals_prob(
  object,
  approximation = c("none", "gpd", "empirical", "rank"),
  threshold_probability = 0.1,
  ...
)
}
\arguments{
\item{object}{A model or numerical data set}

\item{...}{Other arguments are passed to the appropriate method.}

\item{approximation}{Character string specifying the method to use in
computing the surprisal probabilities. See Details below.}

\item{threshold_probability}{Probability threshold when computing the GPD
approximation. This is the probability below which the GPD is fitted. Only
used if \code{approximation = "gpd"}.}
}
\value{
A numerical vector containing the surprisals or surprisal probabilities.
}
\description{
A surprisal is given by \eqn{s = -\log f(y)} where \eqn{f} is the
density or probability mass function of the estimated or assumed distribution,
and \eqn{y} is an observation. This is returned by \code{surprisals()}.
A surprisal probability is the probability of a surprisal at least as extreme
as \eqn{s}. This is returned by \code{surprisals_prob()}
}
\details{
The surprisal probabilities may be computed in three different ways.
\enumerate{
\item When \code{approximation = "none"} (the default), the surprisal probabilities are computed
using the same distribution that was used to compute the surprisal values.
Under this option, surprisal probabilities are equal to 1 minus the
coverage probability of the largest HDR that contains each value. Surprisal
probabilities smaller than 1e-6 are returned as 1e-6.
\item When \code{approximation = "gdp"}, the surprisal probabilities are
computed using a Generalized Pareto Distribution fitted to the most extreme
surprisal values (those with probability less than \code{threshold_probability}).
For surprisal probabilities greater than \code{threshold_probability}, the value of
\code{threshold_probability} is returned. Under this option, the distribution is
used for computing the surprisal values but not for determining their
probabilities. Due to extreme value theory, the resulting probabilities should
be relatively insensitive to the distribution used in computing the surprisal
values.
\item When \code{approximation = "empirical"} (or \code{"rank"}), the surprisal probability of each
observation is estimated using the proportion of observations with
greater or equal surprisal values. This is a nonparametric approach that is also
insensitive to the distribution used in computing the surprisal values.
}
}
\references{
Hyndman, R J (2026) "That's weird: Anomaly detection using R",
Chapter 7, \url{https://OTexts.com/weird/density.html}.

Hyndman, R J & Frazier, D T (2026) "Anomaly detection using surprisals",
\url{https://robjhyndman.com/publications/surprisals.html}.
}
\seealso{
For specific methods, see \code{\link[=surprisals.numeric]{surprisals.numeric()}} and \code{\link[=surprisals.lm]{surprisals.lm()}},
}
\author{
Rob J Hyndman
}
