% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{us_mortality}
\alias{us_mortality}
\title{US weekly mortality}
\format{
A data frame with 10314 rows and
6 columns:
\describe{
\item{Year}{Year}
\item{Week}{ISO week of the year}
\item{Sex}{"Female", "Male" or "Total"}
\item{Age}{Age group: "0-14", "15-64", "65-74", "75-84", "85+" or "Total"}
\item{Deaths}{Number of deaths}
\item{Mortality}{Weekly death rate: deaths divided by population exposure}
}
}
\source{
Human Mortality Database. Max Planck Institute for Demographic
Research (Germany), University of California, Berkeley (USA), and French
Institute for Demographic Studies (France). \url{https://www.mortality.org}
}
\usage{
us_mortality
}
\value{
Data frame
}
\description{
Weekly deaths and death rates in the USA, by sex and age group, from the
second week of 2015 to the last week of 2025. The data are from the
Short-Term Mortality Fluctuations (STMF) series of the Human Mortality
Database.
}
\examples{
us_mortality |>
  filter(Sex == "Total", Age != "Total") |>
  mutate(time = Year + (Week - 1) / 52) |>
  ggplot(aes(x = time, y = Mortality, colour = Age)) +
  geom_line() +
  scale_y_log10()
}
\references{
Hyndman, R J (2026) "That's weird: Anomaly detection using R", Chapter 10,
\url{https://OTexts.com/weird/}.
}
\keyword{datasets}
