% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conflicts.R
\name{weird_conflicts}
\alias{weird_conflicts}
\title{Conflicts between weird packages and other packages}
\usage{
weird_conflicts()
}
\value{
A list object of class \code{weird_conflicts}.
}
\description{
This function lists all the conflicts between packages in the weird collection
and other packages that you have loaded.
}
\details{
Some conflicts are deliberately ignored: \code{intersect}, \code{union},
\code{setequal}, and \code{setdiff} from dplyr; and \code{intersect},
\code{union}, \code{setdiff}, and \code{as.difftime} from lubridate.
These functions make the base equivalents generic, so shouldn't negatively affect any
existing code.
}
\examples{
weird_conflicts()
}
