% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{weird_packages}
\alias{weird_packages}
\title{List all packages loaded by weird}
\usage{
weird_packages(include_self = FALSE)
}
\arguments{
\item{include_self}{Include weird in the list?}
}
\value{
A character vector of package names.
}
\description{
List all packages loaded by weird
}
\examples{
weird_packages()
}
