% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{wine_reviews}
\alias{wine_reviews}
\title{Wine prices and points}
\format{
A data frame with 111,593 rows and 10 columns:
\describe{
\item{country}{Country of origin}
\item{state}{State or province of origin}
\item{region}{Region of origin}
\item{winery}{Name of vineyard that made the wine}
\item{variety}{Variety of grape}
\item{title}{Title of wine review}
\item{description}{A review of the wine}
\item{points}{Points allocated by WineEnthusiast reviewer on a scale of 0-100}
\item{price}{Price of a bottle of wine in $US}
\item{year}{Year of wine extracted from \code{title}}
}
}
\source{
\url{http://kaggle.com}
}
\usage{
wine_reviews
}
\description{
A data set containing data on wines from 44 countries, taken from \emph{Wine Enthusiast Magazine}
during the week of 15 June 2017.
}
\keyword{datasets}
