% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{autoplot.kde}
\alias{autoplot.kde}
\title{Produce ggplot of densities in 1 or 2 dimensions}
\usage{
\method{autoplot}{kde}(
  object,
  prob = seq(9)/10,
  fill = FALSE,
  show_hdr = FALSE,
  show_points = FALSE,
  color = "#00659e",
  palette = hdr_palette,
  alpha = min(1, 1000/NROW(object$x)),
  ...
)
}
\arguments{
\item{object}{Probability density function as estimated by \code{ks::kde()}.}

\item{prob}{Probability of the HDR contours to be drawn (for a bivariate plot only).}

\item{fill}{If \code{TRUE}, and the density is bivariate, the bivariate contours
are shown as filled regions rather than lines.}

\item{show_hdr}{If \code{TRUE}, and the density is univariate, then the HDR regions
specified by \code{prob} are shown as a ribbon below the density.}

\item{show_points}{If \code{TRUE}, then individual points are plotted.}

\item{color}{Color to use for HDR contours (if \code{fill} is \code{FALSE}), or the central color in HDR regions}

\item{palette}{Color palette function to use for HDR filled regions (if \code{fill} is \code{TRUE}).}

\item{alpha}{Transparency of points. Defaults to min(1, 1000/n), where n is the number of observations.}

\item{...}{Additional arguments are currently ignored.}
}
\description{
Produce ggplot of densities in 1 or 2 dimensions
}
\examples{
# Univariate density
c(rnorm(500), rnorm(500, 4, 1.5)) |>
  kde() |>
  autoplot(show_hdr = TRUE, prob= c(0.5, 0.95), color = "#c14b14")
ymat <- tibble(y1 = rnorm(5000), y2 = y1 + rnorm(5000))
ymat |>
  kde(H = Hns(ymat)) |>
  autoplot(fill = TRUE, prob = c(0.5, 0.95), show_points = TRUE)
}
