% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{autoplot.kde}
\alias{autoplot.kde}
\title{Produce ggplot of densities from kde() in 1 or 2 dimensions}
\usage{
\method{autoplot}{kde}(
  object,
  prob = seq(9)/10,
  filled = FALSE,
  show_hdr = FALSE,
  palette = viridisLite::viridis,
  ...
)
}
\arguments{
\item{object}{Probability density function as estimated by \code{ks::kde()}.}

\item{prob}{Probability of the HDR contours to be drawn (for a bivariate plot only).}

\item{filled}{If \code{TRUE}, the bivariate contours are shown as shaded regions rather than lines.}

\item{show_hdr}{If \code{TRUE}, the HDR regions specified by \code{prob} are shown as a ribbon below the density.}

\item{palette}{Color palette to use for HDR regions.}

\item{...}{Additional arguments are ignored.}
}
\description{
Produce ggplot of densities from kde() in 1 or 2 dimensions
}
\examples{
# Univariate density
c(rnorm(100), rnorm(100, 3, 1)) |>
  ks::kde() |>
  autoplot(show_hdr = TRUE, prob= c(0.5, 0.99)) +
  scale_fill_brewer(palette = "OrRd", direction = -1)
tibble(y1 = rnorm(100), y2 = y1 + rnorm(100)) |>
  ks::kde(H = 0.4*diag(2)) |>
  autoplot(filled = TRUE)
}
