% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_df.R
\name{density_df}
\alias{density_df}
\title{Convert distributional object to a data frame}
\usage{
density_df(object)
}
\arguments{
\item{object}{A distributional object}
}
\value{
Data frame with columns \code{x}, \code{y} (if bivariate), \code{density}, and \code{distribution}.
}
\description{
Make a long-form data frame containing densities from a distributional
object on a regular grid for plotting.
}
\examples{
dist_kde(oldfaithful$duration) |> density_df()

}
