% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_kde.R
\name{dist_kde}
\alias{dist_kde}
\title{Create distributional object based on a kernel density estimate}
\usage{
dist_kde(y, h = NULL, H = NULL, lookout = FALSE, multiplier = 1, ...)
}
\arguments{
\item{y}{Numerical vector or matrix of data, or a list of such objects. If a
list is provided, then all objects should be of the same dimension. e.g.,
all vectors, or all matrices with the same number of columns.}

\item{h}{Bandwidth for univariate distribution. If \code{NULL}, the
\code{\link{kde_bandwidth}} function is used.}

\item{H}{Bandwidth matrix for multivariate distribution. If \code{NULL}, the
\code{\link{kde_bandwidth}} function is used.}

\item{lookout}{A logical variable passed to \code{\link[=kde_bandwidth]{kde_bandwidth()}} (set to
\verb{FALSE`` by default) indicating which bandwidth estimator to use. Ignored if }h\code{or}H` are specified.}

\item{multiplier}{Multiplier for bandwidth passed to \code{\link{kde_bandwidth}}.
Ignored if \code{h} or \code{H} are specified.}

\item{...}{Other arguments are passed to \code{\link[ks]{kde}}.}
}
\description{
Creates a distributional object using a kernel density estimate with a
Gaussian kernel obtained from the \code{\link[ks]{kde}()} function. The bandwidth
can be specified; otherwise the \code{\link{kde_bandwidth}()} function is used.
The cdf, quantiles and moments are consistent with the kde. Generating
random values from the kde is equivalent to a smoothed bootstrap.
}
\examples{
dist_kde(c(rnorm(200), rnorm(100, 5)), multiplier = 2)
dist_kde(cbind(rnorm(200), rnorm(200, 5)))

}
