% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{dist_kde}
\alias{dist_kde}
\title{Create kde distributional object}
\usage{
dist_kde(y, h = NULL, H = NULL, kde_options = NULL, ...)
}
\arguments{
\item{y}{Numerical vector or matrix of data, or a list of such objects. If a
list is provided, then all objects should be of the same dimension. e.g.,
all vectors, or all matrices with the same number of columns.}

\item{h}{Bandwidth for univariate distribution. If \code{NULL}, the
\code{\link{kde_bandwidth}} function is used.}

\item{H}{Bandwidth matrix for multivariate distribution. If \code{NULL}, the
\code{\link{kde_bandwidth}} function is used.}

\item{kde_options}{A named list containing arguments for \code{\link[ks]{kde}}.}

\item{...}{Other arguments are passed to the \code{\link{kde_bandwidth}} function.}
}
\description{
Creates a distributional object using a kernel density estimate. The
\code{ks::kde()} function is used to compute the density. The cdf, quantiles and
moments are consistent with the kde. Generating random values from the kde
is equivalent to a smoothed bootstrap.
}
\examples{
dist_kde(c(rnorm(200), rnorm(100, 5)), method = "double")

}
