% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmm.R
\name{dist_mclust}
\alias{dist_mclust}
\title{Convert Gaussian mixture model to a distributional object}
\usage{
dist_mclust(object)
}
\arguments{
\item{object}{Object of class \code{Mclust}, output from the [\link[mclust:Mclust]{mclust::Mclust}] function}
}
\value{
An object of class \code{distributional}
}
\description{
Convert Gaussian mixture model to a distributional object
}
\examples{
library(mclust)
# Univariate mixture density
gmm <- Mclust(oldfaithful$duration) |> dist_mclust()
gg_density(gmm) +
  geom_jitter(data = oldfaithful, aes(x=duration, y = -0.0002),
    width=0, height=0.0002, alpha = 0.1) +
  labs(x = "Eruption duration", y = "")
}
