% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdr.R
\name{hdr_table}
\alias{hdr_table}
\title{Table of Highest Density Regions}
\usage{
hdr_table(y = NULL, density = NULL, prob = c(0.5, 0.99), ...)
}
\arguments{
\item{y}{Numerical vector or matrix of data}

\item{density}{Probablity density function estimated by \code{ks::kde()}.}

\item{prob}{Probability of the HDR}

\item{...}{If \code{y} is supplied, other arguments are passed to \code{\link[ks]{kde}}.
Otherwise, additional arguments are ignored.}
}
\value{
A tibble
}
\description{
Compute the highest density regions (HDR) for a kernel density estimate. The HDRs
are returned as a tibble with one row per interval and columns:
\code{prob} (giving the probability coverage),
\code{density} (the value of the density at the boundary of the HDR),
\code{mode} (the local mode equal to the highest density value within the corresponding HDR).
For one dimensional density functions, the tibble also has columns
\code{lower} (the lower ends of the intervals), and
\code{upper} (the upper ends of the interval).
}
\examples{
hdr_table(density = dnorm, prob = c(0.9, 0.95))
hdr_table(c(rnorm(50), rnorm(50, 3, 1)), prob = 0.95)
hdr_table(c(rnorm(50), rnorm(50, 3, 1)), prob = 0.95)
}
\author{
Rob J Hyndman
}
