% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdr.R
\name{hdr_table}
\alias{hdr_table}
\title{Table of Highest Density Regions}
\usage{
hdr_table(x = NULL, y = NULL, density = NULL, prob = c(50, 99), bw = NULL, ...)
}
\arguments{
\item{x}{Numerical vector of data}

\item{y}{Numerical vector of data}

\item{density}{Density estimate}

\item{prob}{Probability of the HDR}

\item{bw}{Bandwidth for density estimate}
}
\value{
A tibble with columns prob, lower and upper
}
\description{
Compute the highest density regions (HDR) for either a specified density function, or
a kernel density estimate based on the data provided. Either 1 or 2 dimensional HDRs can be
calculated. The HDRs are returned as a tibble with one row per interval and columns:
\code{prob} (giving the probability coverage),
\code{lower} (the lower end of the interval),
\code{upper} (the upper end of the interval), and
\code{density} (the value of the density at the ends of the interval).
The first row of the tibble has \code{prob} set to 0 and specifies the mode of the distribution.
}
\examples{
x <- c(rnorm(49), 5)
hdr_table(x, prob = 0.95)
}
\author{
Rob J Hyndman
}
