% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{kde_bandwidth}
\alias{kde_bandwidth}
\title{Robust bandwidth estimation for kernel density estimation}
\usage{
kde_bandwidth(data)
}
\arguments{
\item{data}{A numeric matrix or data frame.}
}
\value{
A matrix of bandwidths (or scalar in the case of univariate data).
}
\description{
Robust bandwidth estimation for kernel density estimation
}
\author{
Rob J Hyndman
}
