% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde_bandwidth.R
\name{kde_bandwidth}
\alias{kde_bandwidth}
\title{Robust bandwidth estimation for kernel density estimation}
\usage{
kde_bandwidth(data, multiplier = 1)
}
\arguments{
\item{data}{A numeric matrix or data frame.}

\item{multiplier}{Bandwidths are chosen using a robust version of the
normal reference rule multiplied by a constant. The default is 1.}
}
\value{
A matrix of bandwidths (or scalar in the case of univariate data).
}
\description{
Robust bandwidth estimation for kernel density estimation
}
\examples{
# Univariate bandwidth calculation
kde_bandwidth(oldfaithful$duration)
# Bivariate bandwidth calculation
kde_bandwidth(oldfaithful[, c("duration", "waiting")])
}
\author{
Rob J Hyndman
}
