% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde_bandwidth.R
\name{kde_bandwidth}
\alias{kde_bandwidth}
\title{Robust bandwidth estimation for kernel density estimation}
\usage{
kde_bandwidth(data, method = c("rnrr", "lookout"), ...)
}
\arguments{
\item{data}{A numeric matrix or data frame.}

\item{method}{A character string giving the method to use. If equal to \code{rnrr},
a robust version of the normal reference rule is used. If equal to \code{lookout},
the bandwidth matrix estimate of Hyndman, Kandanaarachchi & Turner (2026)
is returned.}

\item{...}{Additional arguments are ignored if \code{method = "lookout"} and passed to
\code{\link[lookout:find_tda_bw]{lookout::find_tda_bw()}} otherwise.}
}
\value{
A matrix of bandwidths (or a scalar in the case of univariate data).
}
\description{
Bandwidth matrices are estimated using either a robust version of the normal
reference rule, or using the approach of Hyndman, Kandanaarachchi & Turner (2026).
}
\examples{
# Univariate bandwidth calculation
kde_bandwidth(oldfaithful$duration)
# Bivariate bandwidth calculation
kde_bandwidth(oldfaithful[, c("duration", "waiting")])
}
\references{
Rob J Hyndman, Sevvandi Kandanaarachchi & Katharine Turner (2026)
"When lookout sees crackle: Anomaly detection via kernel density estimation",
unpublished. \url{https://robjhyndman.com/publications/lookout2/}

Rob J Hyndman (2026) "That's weird: Anomaly detection using R",
\url{https://OTexts.com/weird/}
}
\author{
Rob J Hyndman
}
