% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde_scores.R
\name{kde_scores}
\alias{kde_scores}
\title{Kernel density estimate scores and lookout probabilities}
\usage{
kde_scores(y, loo = FALSE, h = kde_bandwidth(y), H = kde_bandwidth(y), ...)
}
\arguments{
\item{y}{Numerical vector or matrix of data (up to 6 dimensions).}

\item{loo}{Should leave-one-scores be returned? Default: FALSE.}

\item{h}{Bandwidth for univariate kernel density estimate. Default is \code{\link{kde_bandwidth}}.}

\item{H}{Bandwidth for multivariate kernel density estimate. Default is \code{\link{kde_bandwidth}}.}

\item{...}{Other arguments are passed to \code{\link[ks]{kde}}.}
}
\value{
Numerical vector containing kde scores.
}
\description{
Compute kernel density estimate scores,
defined as minus the log of the kernel density estimate at each observation.
A Gaussian kernel is used, with default bandwidth given by a normal reference rule.
Leave-one-out scores are optionally computed.
Lookout probabilities are returned by \code{lookout_prob} computed using a
Generalized Pareto Distribution estimated from the kde scores and applied
to the leave-one-out kde scores. These give the probability of each observation
being an anomaly.
}
\examples{
# Univariate data
y <- c(5, rnorm(49))
kde_scores(y)
lookout_prob(y)
# Bivariate data
y <- cbind(rnorm(50), c(5, rnorm(49)))
kde_scores(y)
lookout_prob(y)
}
\seealso{
\code{\link{kde_bandwidth}}
\code{\link[ks]{kde}}
}
\author{
Rob J Hyndman
}
