% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde_scores.R
\name{kde_scores}
\alias{kde_scores}
\alias{lookout_prob}
\title{Kernel density estimate scores and lookout probabilities}
\usage{
kde_scores(y, loo = FALSE, h = stats::bw.nrd0(y), ...)

lookout_prob(y, h = stats::bw.nrd0(y), ...)
}
\arguments{
\item{y}{Numerical vector or matrix of data (up to 6 dimensions).}

\item{loo}{Should leave-one-scores be returned? Default: FALSE.}

\item{h}{Bandwidth for the kernel density estimate. Default: normal reference rule given
by \code{\link[stats]{bw.nrd0}}.}

\item{...}{Other arguments are passed to \code{\link[ks]{kde}}.}
}
\value{
Numerical vector containing kde scores.
}
\description{
Compute kernel density estimate scores,
defined as minus the log of the kernel density estimate at each observation.
A Gaussian kernel is used, with default h given by a normal reference rule.
Leave-one-out scores are optionally computed.
Lookout probabilities are returned by \code{lookout_prob} computed using a
Generalized Pareto Distribution estimated from the kde scores and applied
to the leave-one-out kde scores. These give the probability of each observation
being an anomaly.
}
\examples{
y <- c(rnorm(49), 5)
kde_scores(y)
lookout_prob(y)
}
\seealso{
\code{\link[stats]{bandwidth}}
\code{\link[ks]{kde}}
}
\author{
Rob J Hyndman
}
