% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde_scores.R
\name{kde_scores}
\alias{kde_scores}
\alias{lookout_prob}
\title{Kernel density estimate scores and lookout probabilities}
\usage{
kde_scores(
  y,
  h = stats::bw.nrd(y),
  kernel = c("gaussian", "epanechnikov"),
  loo = FALSE
)

lookout_prob(y, h = stats::bw.nrd(y), kernel = c("gaussian", "epanechnikov"))
}
\arguments{
\item{y}{Numerical vector of data}

\item{h}{Bandwidth. Default: \code{\link[stats]{bw.nrd}(y)}}

\item{kernel}{A character string giving the kernel to be used.}

\item{loo}{Should leave-one-scores be returned? Default: FALSE}
}
\value{
Numerical vector containing kde scores
}
\description{
Compute kernel density estimate scores,
defined as minus the log of the kernel density estimate at each observation.
A Gaussian kernel is used, with default h given by a normal reference rule.
Leave-one-out scores are optionally computed.
Lookout probabilities are returned by \code{lookout_prob} computed using a
Generalized Pareto Distribution estimated from the kde scores and applied
to the leave-one-out kde scores. These give the probability of each observation
being an anomaly.
}
\examples{
y <- c(rnorm(49), 5)
kde_scores(y)
lookout_prob(y)
}
\seealso{
\code{\link[stats]{bandwidth}}
\code{\link[ks]{kde}}
}
\author{
Rob J Hyndman
}
