% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_scores.R
\name{reg_scores}
\alias{reg_scores}
\title{Regression scores from a linear model}
\usage{
reg_scores(object, loo = FALSE)
}
\arguments{
\item{object}{A fitted model, usually produced by \code{\link[stats]{lm}}.}

\item{loo}{Should leave-one-scores be returned? Default: FALSE.}
}
\value{
Numerical vector containing log scores.
}
\description{
Compute log scores and log-loo scores from a linear Gaussian regression model,
defined as minus the log of the conditional density computed at each observation.
Leave-one-out scores are optionally computed.
}
\author{
Rob J Hyndman
}
