% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_tail_prob.R
\name{score_tail_prob}
\alias{score_tail_prob}
\title{Density score tail probabilities}
\usage{
score_tail_prob(g, distribution = NULL, n = 1e+06)
}
\arguments{
\item{g}{vector of density scores.}

\item{distribution}{A distributional object specifying the distribution to
use.}

\item{n}{Number of samples to generate for estimating the probabilities.}
}
\description{
Compute the probability of a density score at least as extreme as \code{g} given
a specified distribution. A density score is given by $g = -\log(f)$ where
$f$ is the density or probability mass function of the distribution. The
argument \code{distribution} specifies which distribution to use. The
probabilities are estimate by simulation; \code{n} random values are generated
from the specified \code{distribution}. Consequently, slightly different values
will be returned each time unless a random seed is set using
\link{\code[base]{set.seed}}.
}
\examples{
library(distributional)
score_tail_prob(-log(dnorm(1:4)), dist_normal())
}
