% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surprisal_prob.R
\name{surprisal_prob}
\alias{surprisal_prob}
\title{Surprisal probabilities}
\usage{
surprisal_prob(
  object,
  distribution = NULL,
  loo = FALSE,
  GPD = FALSE,
  threshold_probability = 0.1,
  ...
)
}
\arguments{
\item{object}{A model or numerical data set.}

\item{distribution}{A probability distribution stored as a distributional
object. Ignored if \code{object} is a model.}

\item{loo}{Logical value specifying if leave-one-out surprisals should be computed.}

\item{GPD}{Logical value specifying if a Generalized Pareto distribution
should be used to estimate the probabilities.}

\item{threshold_probability}{Probability threshold when computing the GPD
distribution for the surprisals.}

\item{...}{Other arguments are passed to \code{\link{surprisals}}.}
}
\description{
Compute the probability of a surprisal at least as extreme as those observed.
A surprisal is given by \eqn{-\log(f)} where \eqn{f} is the density or
probability mass function of the distribution.
The surprisal values are computed from the \code{distribution} provided. If no
distribution is provided, a kernel density estimate is used.
}
\details{
The surprisal probabilities may be computed in three different ways.
\enumerate{
\item Given the same \code{distribution} that was used to compute the surprisal values.
Under this option, surprisal probabilities are equal to 1 minus the
coverage probability of the largest HDR that contains each value. Surprisal
probabilities smaller than 1e-6 are returned as 1e-6.
\item Using a Generalized Pareto Distribution fitted to the most extreme
surprisal values (those with probability less than \code{threshold_probability}).
This option is used if \code{GPD = TRUE}. For surprisal values with
probability less than \code{threshold_probability}, the value of
\code{threshold_probability} is returned. Under this option, the distribution is
used for computing the surprisal values but not for determining their
probabilities. Due to extreme value theory, the resulting probabilities should
be relatively insensitive to the distribution used in computing the surprisal
values.
\item Empirically as the proportion of observations with greater surprisal values.
This option is used when \code{GPD = FALSE} and no distribution is explicitly
provided. This is also insensitive to the distribution used in computing the
surprisal values.
}
}
\examples{
# Univariate data
tibble(
  y = c(5, rnorm(49)),
  p = surprisal_prob(y)
)
tibble(
  y = n01$v1,
  prob1 = surprisal_prob(y),
  prob2 = surprisal_prob(y, GPD = TRUE),
  prob3 = surprisal_prob(y, dist_normal()),
  prob4 = surprisal_prob(y, dist_normal(), GPD = TRUE)
) |>
  arrange(prob1)
# Bivariate data
tibble(
  x = rnorm(50),
  y = c(5, rnorm(49)),
  lookout = lookout_prob(cbind(x, y))
)
# Using a regression model
of <- oldfaithful |> filter(duration < 7200, waiting < 7200)
fit_of <- lm(waiting ~ duration, data = of)
of |>
  mutate(p = surprisal_prob(fit_of)) |>
  arrange(p)
}
