% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surprisals.R
\name{surprisals.default}
\alias{surprisals.default}
\title{Surprisals computed from a data set}
\usage{
\method{surprisals}{default}(
  object,
  distribution = dist_kde(object, multiplier = 2, ...),
  loo = FALSE,
  ...
)
}
\arguments{
\item{object}{A numerical data  set, either a vector, matrix or data frame.}

\item{distribution}{A probability distribution stored as a distributional
object.}

\item{loo}{Should leave-one-out surprisals be computed?}

\item{...}{Other arguments are passed to \code{\link{dist_kde}}.}
}
\value{
A numerical vector containing the surprisals.
}
\description{
Compute surprisals from a data set given a probability
distribution. The surprisals are defined as minus the log of the density at
each observation.
}
\details{
If no distribution is provided, a kernel density estimate is
computed. The leave-one-out surprisals (or LOO surprisals) are obtained by
estimating the kernel density estimate using all other observations.
}
\examples{
# surprisals computed from bivariate data set
oldfaithful |>
  filter(duration < 7000, waiting < 7000) |>
  mutate(
    loo_fscores = surprisals(cbind(duration, waiting), loo = TRUE)
  )
}
\seealso{
\code{\link{dist_kde}}
}
\author{
Rob J Hyndman
}
