% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surprisals.R
\name{surprisals}
\alias{surprisals}
\title{Surprisals}
\usage{
surprisals(object, ...)
}
\arguments{
\item{object}{A model or numerical data set}

\item{...}{Other arguments are passed to the appropriate method.}
}
\description{
Compute surprisals from a model or a data set given a
probability distribution. The surprisals are defined as minus the log of the
density at each observation.
}
