% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surprisals.R
\name{surprisals.lm}
\alias{surprisals.lm}
\alias{surprisals.gam}
\title{Surprisals computed from a model}
\usage{
\method{surprisals}{lm}(object, loo = FALSE, ...)

\method{surprisals}{gam}(object, loo = FALSE, ...)
}
\arguments{
\item{object}{A model object such as returned by \code{\link[stats]{lm}},
or \code{\link[mgcv]{gam}}.}

\item{loo}{Should leave-one-out surprisals be computed?}

\item{...}{Other arguments are ignored.}
}
\description{
Surprisals computed from a model
}
\examples{
# surprisals computed from linear model
of <- oldfaithful |>
  filter(duration < 7200, waiting < 7200)
lm_of <- lm(waiting ~ duration, data = of)
of |>
  mutate(
    fscore = surprisals(lm_of),
    loo_fscore = surprisals(lm_of, loo = TRUE),
    #lookout_prob = lookout(surprisals = fscore, loo_scores = loo_fscore)
  ) |>
  ggplot(aes(x = duration, y = waiting,
    color = loo_fscore > quantile(loo_fscore, 0.99))) +
  geom_point()
# surprisals computed from GAM
of <- oldfaithful |>
  filter(duration > 1, duration < 7200, waiting < 7200)
gam_of <- mgcv::gam(waiting ~ s(duration), data = of)
of |>
  mutate(fscore = surprisals(gam_of))
}
