% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weird.R
\docType{package}
\name{weird-package}
\alias{weird}
\alias{weird-package}
\title{weird: Functions and Data Sets for "That's Weird: Anomaly Detection Using R" by Rob J Hyndman}
\description{
All functions and data sets required for the examples in the book Hyndman (2024) "That's Weird: Anomaly Detection Using R" \url{https://OTexts.com/weird/}. All packages needed to run the examples are also loaded.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pkg.robjhyndman.com/weird-package/}
  \item \url{https://github.com/robjhyndman/weird-package}
  \item Report bugs at \url{https://github.com/robjhyndman/weird-package/issues}
}

}
\author{
\strong{Maintainer}: Rob Hyndman \email{Rob.Hyndman@monash.edu} (\href{https://orcid.org/0000-0002-2140-5352}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
\keyword{internal}
